Chore: extract duplicated code into a superclass.
